+++ test/test_scrabble_game.py
@@ -248,5 +248,6 @@ class TestCalculateWordValue(unittest.TestCase):
     def test_player_point_adding_letter_to_existing_word(self):
         pass
 
+
 if __name__ == '__main__':
     unittest.main()